#40 Rename: common to `global`
