+++ src/test/java/com/apps/pochak/block/controller/BlockControllerTest.java
@@ -20,8 +20,8 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.filter.CharacterEncodingFilter;
 
-import static com.apps.pochak.common.ApiDocumentUtils.getDocumentRequest;
-import static com.apps.pochak.common.ApiDocumentUtils.getDocumentResponse;
+import static com.apps.pochak.global.ApiDocumentUtils.getDocumentRequest;
+import static com.apps.pochak.global.ApiDocumentUtils.getDocumentResponse;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
